Control coalescing type through flag
